fix(match): move conflate class to separate file
